dm: sync agk's latest non-thinp DM changes from his editing tree (for 3.1)

+++ drivers/md/dm-flakey.c
@@ -255,10 +255,10 @@ static void corrupt_bio_data(struct bio *bio, struct flakey_c *fc)
 		data[fc->corrupt_bio_byte - 1] = fc->corrupt_bio_value;
 
 		DMDEBUG("Corrupting data bio=%p by writing %u to byte %u "
-			"(rw=%c bi_rw=%lu bi_sector=%lu cur_bytes=%u)\n",
+			"(rw=%c bi_rw=%lu bi_sector=%llu cur_bytes=%u)\n",
 			bio, fc->corrupt_bio_value, fc->corrupt_bio_byte,
 			(bio_data_dir(bio) == WRITE) ? 'w' : 'r',
-			bio->bi_rw, bio->bi_sector, bio_bytes);
+			bio->bi_rw, (unsigned long long)bio->bi_sector, bio_bytes);
 	}
 }
